Addition, subtraction and multiplication
Implement those three operations and division with some errors. What's going to be done next is to modify the algorithm in an attempt to get it to work. However, since it may be a bit change, it's preferable to commit all that's been done until now in case I mess up.